Refactored to database.py
